useDocs hook の作成
